v0.8.0  Break in compatibility: DType int64 -> Go's int64 (it was Go's int)  and many small improvements.

+++ examples/cifar/cifar.go
@@ -130,7 +130,7 @@ func LoadCifar10(manager *Manager, baseDir string, dtype shapes.DType) (partitio
 	labels := tensor.FromShape(shapes.Make(shapes.I64, NumExamples, 1))
 	labelsRef := labels.AcquireData()
 	defer labelsRef.Release()
-	labelsData := labelsRef.Flat().([]int)
+	labelsData := labelsRef.Flat().([]int64)
 
 	var labelImageBytes [imageSizeBytes + 1]byte
 	for fileIdx := 0; fileIdx < 6; fileIdx++ {
@@ -165,7 +165,7 @@ func LoadCifar10(manager *Manager, baseDir string, dtype shapes.DType) (partitio
 			if err != nil {
 				panic(errors.WithMessagef(err, "failed converting bytes to tensor of %s", dtype))
 			}
-			labelsData[exampleIdx] = int(labelImageBytes[0])
+			labelsData[exampleIdx] = int64(labelImageBytes[0])
 		}
 	}
 	return partitionImagesAndLabels(manager, images, labels)
@@ -184,7 +184,7 @@ func LoadCifar100(manager *Manager, baseDir string, dtype shapes.DType) (partiti
 	labels := tensor.FromShape(shapes.Make(shapes.I64, NumExamples, 1))
 	labelsRef := labels.AcquireData()
 	defer labelsRef.Release()
-	labelsData := labelsRef.Flat().([]int)
+	labelsData := labelsRef.Flat().([]int64)
 
 	var labelImageBytes [imageSizeBytes + 2]byte
 	dataFiles := []string{"train.bin", "test.bin"}
@@ -218,7 +218,7 @@ func LoadCifar100(manager *Manager, baseDir string, dtype shapes.DType) (partiti
 			if err != nil {
 				panic(errors.Wrapf(err, "failed converting bytes to tensor of %s", dtype))
 			}
-			labelsData[exampleIdx] = int(labelImageBytes[1]) // Take the fine-label (and discard the coarse-label).
+			labelsData[exampleIdx] = int64(labelImageBytes[1]) // Take the fine-label (and discard the coarse-label).
 			exampleIdx++
 		}
 	}
